Add Custom entry file
