+++ .changeset/pink-peaches-wave.md
@@ -1,0 +1,6 @@
+---
+"svelte-adapter-fastly": patch
+---
+
+Add Ability to supply custom entry files
+Removed unnecessary build step.
